requirements.txt
altgraph==0.16.1
cycler==0.10.0
future==0.18.2
kiwisolver==1.1.0
matplotlib==3.1.2
nibabel==2.5.1
numpy==1.17.4
pefile==2019.4.18
Pillow==6.2.1
PyInstaller==3.5
pyparsing==2.4.5
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
six==1.13.0
wxPython==4.0.7.post2
